Improve/fix du and cp commands for non-Linux platforms
- On Windows, robocopy is much more performant at copying things around
  and can also be used as a 'du' replacement. Now also there is less
  need for cygwin/Gnuwin frameworks.
- The way du was being called was generally also broken.
